requirements.txt
opencv-python
pytesseract
#tesserocr
Pillow